requirements.txt
torch>=1.10.0
torchvision>=0.11.0
facenet-pytorch>=2.5.2
numpy>=1.19.0
pandas>=1.1.0
scikit-learn>=1.0.0
opencv-python>=4.5.0
Pillow>=9.0.0
flask>=2.0.0
joblib>=1.0.0
matplotlib>=3.3.0
tqdm>=4.60.0